Use BigDecimal

+++ src/main/java/com/zopa/quote/ILoanValidator.java
@@ -2,9 +2,10 @@ package com.zopa.quote;
 
 import com.zopa.exception.QuoteNotPossibleException;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @FunctionalInterface
 public interface ILoanValidator<L> {
-    abstract boolean validateLoanProvision(List<L> lenders, long loanAmount) throws QuoteNotPossibleException;
+    abstract boolean validateLoanProvision(List<L> lenders, BigDecimal loanAmount) throws QuoteNotPossibleException;
 }
