+++ src/main/java/com/zopa/runner/DefaultQuoteRunner.java
@@ -13,6 +13,7 @@ import com.zopa.quote.IQuoteCalculator;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor
@@ -20,7 +21,7 @@ import java.util.List;
 public class DefaultQuoteRunner implements IQuoteRunner {
 
     private String fileName;
-    private Long loanAmount;
+    private BigDecimal loanAmount;
 
     @Override
     public void run() {
